# [mono] dn_simdhash_ght_compatible optimizations for ARM64