--- a/src/native/containers/dn-simdhash-ght-compatible.c
+++ b/src/native/containers/dn-simdhash-ght-compatible.c
@@ -68,13 +68,13 @@ dn_simdhash_ght_replaced (dn_simdhash_ght_data data, void * old_key, void * new_
 unsigned int
 dn_simdhash_ght_default_hash (const void * key)
 {
-    return (unsigned int)(size_t)key;
+	return (unsigned int)(size_t)key;
 }
 
 int32_t
 dn_simdhash_ght_default_comparer (const void * a, const void * b)
 {
-    return a == b;
+	return a == b;
 }
 
 dn_simdhash_ght_t *
@@ -84,13 +84,13 @@ dn_simdhash_ght_new (
 )
 {
 	dn_simdhash_ght_t *hash = dn_simdhash_new_internal(&DN_SIMDHASH_T_META, DN_SIMDHASH_T_VTABLE, capacity, allocator);
-    // Most users of dn_simdhash_ght are passing a custom comparer, and always doing an indirect call ends up being faster
-    //  than conditionally doing a fast inline check when there's no comparer set. Somewhat counter-intuitive, but true
-    //  on both x64 and arm64. Probably due to the smaller code size and reduced branch predictor pressure.
-    if (!hash_func)
-        hash_func = dn_simdhash_ght_default_hash;
-    if (!key_equal_func)
-        key_equal_func = dn_simdhash_ght_default_comparer;
+	// Most users of dn_simdhash_ght are passing a custom comparer, and always doing an indirect call ends up being faster
+	//  than conditionally doing a fast inline check when there's no comparer set. Somewhat counter-intuitive, but true
+	//  on both x64 and arm64. Probably due to the smaller code size and reduced branch predictor pressure.
+	if (!hash_func)
+		hash_func = dn_simdhash_ght_default_hash;
+	if (!key_equal_func)
+		key_equal_func = dn_simdhash_ght_default_comparer;
 	dn_simdhash_instance_data(dn_simdhash_ght_data, hash).hash_func = hash_func;
 	dn_simdhash_instance_data(dn_simdhash_ght_data, hash).key_equal_func = key_equal_func;
 	return hash;
@@ -104,10 +104,10 @@ dn_simdhash_ght_new_full (
 )
 {
 	dn_simdhash_ght_t *hash = dn_simdhash_new_internal(&DN_SIMDHASH_T_META, DN_SIMDHASH_T_VTABLE, capacity, allocator);
-    if (!hash_func)
-        hash_func = dn_simdhash_ght_default_hash;
-    if (!key_equal_func)
-        key_equal_func = dn_simdhash_ght_default_comparer;
+	if (!hash_func)
+		hash_func = dn_simdhash_ght_default_hash;
+	if (!key_equal_func)
+		key_equal_func = dn_simdhash_ght_default_comparer;
 	dn_simdhash_instance_data(dn_simdhash_ght_data, hash).hash_func = hash_func;
 	dn_simdhash_instance_data(dn_simdhash_ght_data, hash).key_equal_func = key_equal_func;
 	dn_simdhash_instance_data(dn_simdhash_ght_data, hash).key_destroy_func = key_destroy_func;
@@ -155,14 +155,14 @@ dn_simdhash_ght_insert_replace (
 
 void *
 dn_simdhash_ght_get_value_or_default (
-    dn_simdhash_ght_t *hash, void * key
+	dn_simdhash_ght_t *hash, void * key
 )
 {
-    check_self(hash);
-    uint32_t key_hash = DN_SIMDHASH_KEY_HASHER(DN_SIMDHASH_GET_DATA(hash), key);
-    DN_SIMDHASH_VALUE_T *value_ptr = DN_SIMDHASH_FIND_VALUE_INTERNAL(hash, key, key_hash);
-    if (value_ptr)
-        return *value_ptr;
-    else
-        return NULL;
+	check_self(hash);
+	uint32_t key_hash = DN_SIMDHASH_KEY_HASHER(DN_SIMDHASH_GET_DATA(hash), key);
+	DN_SIMDHASH_VALUE_T *value_ptr = DN_SIMDHASH_FIND_VALUE_INTERNAL(hash, key, key_hash);
+	if (value_ptr)
+		return *value_ptr;
+	else
+		return NULL;
 }