--- a/src/native/containers/dn-simdhash-ght-compatible.h
+++ b/src/native/containers/dn-simdhash-ght-compatible.h
@@ -31,7 +31,7 @@ dn_simdhash_ght_insert_replace (
 // faster and simpler to use than dn_simdhash_ght_try_get_value, use it wisely
 void *
 dn_simdhash_ght_get_value_or_default (
-    dn_simdhash_ght_t *hash, void * key
+	dn_simdhash_ght_t *hash, void * key
 );
 
 // compatibility shims for the g_hash_table_ versions in glib.h