--- a/src/native/containers/dn-simdhash-ptr-ptr.c
+++ b/src/native/containers/dn-simdhash-ptr-ptr.c
@@ -11,8 +11,8 @@
 #define DN_SIMDHASH_T dn_simdhash_ptr_ptr
 #define DN_SIMDHASH_KEY_T void *
 #define DN_SIMDHASH_VALUE_T void *
-#define DN_SIMDHASH_KEY_HASHER(hash, key) (MurmurHash3_32_ptr(key, 0))
-#define DN_SIMDHASH_KEY_EQUALS(hash, lhs, rhs) (lhs == rhs)
+#define DN_SIMDHASH_KEY_HASHER(data, key) (MurmurHash3_32_ptr(key, 0))
+#define DN_SIMDHASH_KEY_EQUALS(data, lhs, rhs) (lhs == rhs)
 #if SIZEOF_VOID_P == 8
 #define DN_SIMDHASH_BUCKET_CAPACITY 11
 #else