Depends on https://github.com/dotnet/runtime/pull/113095.

The current version of dn_simdhash_ght_compatible has especially bad performance on arm64 vs x64, so I took some steps to redesign it and address some codegen deficiencies. This has the added benefit of making it a bit faster on x64 as well.

There weren't benchmarks for ght_compatible in the benchmark suite originally, so I added a couple. (I was benchmarking end-to-end originally, which makes it harder to compare across platforms.)

The other variants of simdhash should be unaffected, this was mostly a design defect in the ghashtable compatibility variant.

For the find_random_keys benchmark:
**Improved from 109.7% to 76.84% on x64.**
**Improved from 121.17% to 82.22% on ARM64.**

Before (x64):
```
ght_find_random_keys: Warmed 18 time(s). Running 8192 iterations... 14 step(s): avg 17.507ns min 12.824ns max 21.249ns
dnght_find_random_keys: Warmed 21 time(s). Running 8192 iterations... 13 step(s): avg 19.211ns min 18.790ns max 19.541ns
dn_find_random_keys: Warmed 31 time(s). Running 8192 iterations... 19 step(s): avg 13.353ns min 12.913ns max 13.980ns
```

After (x64):
```
ght_find_random_keys: Warmed 16 time(s). Running 4096 iterations... 23 step(s): avg 21.244ns min 13.822ns max 23.902ns
dnght_find_random_keys: Warmed 24 time(s). Running 8192 iterations... 15 step(s): avg 16.325ns min 15.494ns max 16.779ns
dn_find_random_keys: Warmed 30 time(s). Running 8192 iterations... 18 step(s): avg 13.687ns min 13.139ns max 13.834ns
```

Before (arm64):
```
ght_find_random_keys: Warmed 6 time(s). Running 2048 iterations... 15 step(s): avg 69.279ns min 67.222ns max 70.292ns
dnght_find_random_keys: Warmed 5 time(s). Running 2048 iterations... 12 step(s): avg 83.949ns min 83.271ns max 84.813ns
dn_find_random_keys: Warmed 6 time(s). Running 2048 iterations... 15 step(s): avg 69.026ns min 68.155ns max 69.770ns
```

After (arm64):
```
ght_find_random_keys: Warmed 7 time(s). Running 2048 iterations... 15 step(s): avg 65.820ns min 64.640ns max 69.091ns
dnght_find_random_keys: Warmed 8 time(s). Running 2048 iterations... 19 step(s): avg 54.118ns min 53.717ns max 54.888ns
dn_find_random_keys: Warmed 10 time(s). Running 4096 iterations... 13 step(s): avg 40.598ns min 40.352ns max 40.794ns
```